More refactoring....
Rushed the last commit and caused the build not to compile, it should now, and there are some new improvements to the code (I think they are).

Just doing some work to try to resolve the classpath issues still.

* Currently working on the classloading issues.

Fixed a bunch of references to a command that wasn't updated.

Fixed a couple stupid mistakes.. tests seem to almost run, there is one CNFE which is wierd, but the rest seem to resolve.  Child instances get created and started but present a login prompt which causes those tests to fail...

Came up with a better way to avoid the TCCL issues that are hit with hazelcast, a little tedious, but it can be broken down to be a lot more automatic.  Only one test is running right now to make it easy to get everything fixed, which it almost is.